Moves parallel module to timemachine namespace
* Adds `make grpc` to recompile the GRPC code
* Adds grpcio to dev requirements
* Modifies imports
