+++ Makefile
@@ -15,6 +15,10 @@ build:
 clean:
 	cd $(CPP_DIR) && rm -rf build/ | true
 
+.PHONY: grpc
+grpc:
+	python -m grpc_tools.protoc -I grpc/ --python_out=. --grpc_python_out=. grpc/timemachine/parallel/grpc/service.proto
+
 ci:
 	pre-commit run --all-files --show-diff-on-failure && \
 	export PYTHONPATH=$(MKFILE_DIR) && \
