+++ examples/rbfe_single.py
@@ -16,8 +16,8 @@ from testsystems.relative import hif2a_ligand_pair
 
 from timemachine.ff.handlers.serialize import serialize_handlers
 from timemachine.ff.handlers.nonbonded import AM1CCCHandler, LennardJonesHandler
-from parallel.client import CUDAPoolClient
-from parallel.utils import get_gpu_count
+from timemachine.parallel.client import CUDAPoolClient
+from timemachine.parallel.utils import get_gpu_count
 
 from timemachine.optimize.step import truncated_step
 from timemachine.optimize.utils import flatten_and_unflatten
