+++ examples/validate_relative_binding.py
@@ -35,8 +35,8 @@ from timemachine.fe.frames import endpoint_frames_only, all_frames, no_frames
 
 from timemachine.ff import Forcefield
 from timemachine.ff.handlers.deserialize import deserialize_handlers
-from parallel.client import CUDAPoolClient, GRPCClient
-from parallel.utils import get_gpu_count
+from timemachine.parallel.client import CUDAPoolClient, GRPCClient
+from timemachine.parallel.utils import get_gpu_count
 
 
 from timemachine.training.dataset import Dataset
