+++ setup.cfg
@@ -13,7 +13,7 @@ exclude =
   .git,
   __pycache__,
   ff/params,
-  parallel/grpc
+  timemachine/parallel/grpc
 
 [versioneer]
 VCS = git
