+++ slow_tests/test_rbfe.py
@@ -6,8 +6,8 @@ from timemachine.md import builders
 from timemachine.fe.model import RBFEModel
 from timemachine.fe.free_energy import construct_lambda_schedule
 from testsystems.relative import hif2a_ligand_pair
-from parallel.client import CUDAPoolClient
-from parallel.utils import get_gpu_count
+from timemachine.parallel.client import CUDAPoolClient
+from timemachine.parallel.utils import get_gpu_count
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), "..", "tests", "data")
 NUM_GPUS = get_gpu_count()
