+++ tests/test_estimator.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from timemachine.fe import estimator_abfe
 from timemachine.lib import LangevinIntegrator, potentials, MonteCarloBarostat
-from parallel.client import CUDAPoolClient
+from timemachine.parallel.client import CUDAPoolClient
 from timemachine.md.barostat.utils import get_bond_list, get_group_indices
