+++ timemachine/parallel/worker.py
@@ -7,9 +7,9 @@ import logging
 import pickle
 from concurrent import futures
 
-from parallel.grpc import service_pb2, service_pb2_grpc
-from parallel.utils import get_worker_status
-from parallel.constants import DEFAULT_GRPC_OPTIONS
+from timemachine.parallel.grpc import service_pb2, service_pb2_grpc
+from timemachine.parallel.utils import get_worker_status
+from timemachine.parallel.constants import DEFAULT_GRPC_OPTIONS
 
 import grpc
